+++ breeze-core/src/main/kotlin/com/windea/breezeframework/core/extensions/CollectionExtensions.kt
@@ -6,7 +6,6 @@ package com.windea.breezeframework.core.extensions
 import com.windea.breezeframework.core.annotations.*
 import com.windea.breezeframework.core.domain.text.*
 import java.util.*
-import java.util.concurrent.*
 import kotlin.collections.ArrayList
 import kotlin.collections.LinkedHashMap
 import kotlin.collections.LinkedHashSet