+++ breeze-core/src/main/kotlin/com/windea/breezeframework/core/extensions/NumberExtensions.kt
@@ -4,6 +4,9 @@
 package com.windea.breezeframework.core.extensions
 
 import com.windea.breezeframework.core.annotations.*
+import com.windea.breezeframework.core.domain.text.*
+import java.text.*
+import java.util.*
 import kotlin.math.*
 
 //region common extensions
@@ -61,6 +64,28 @@ fun Int.getDigitNumber(index: Int): Int {
 fun Long.getDigitNumber(index: Int): Long {
 	require(index >= 0) { "Index must be non-negative, but was $index." }
 	return this / 10.positivePow(index) % 10
+}
+
+
+/**根据指定的格式化类型，格式化当前数字。可以指定可选的语言环境。*/
+fun Number.formatBy(type:NumberFormatType,locale:Locale? = null):String{
+	return getNumberFormatInstance(type,locale?:Locale.getDefault(Locale.Category.FORMAT)).format(this)
+}
+
+/**根据指定的格式化类型，格式化当前数字。可以指定可选的语言环境。可以进行额外的配置。*/
+fun Number.formatBy(type:NumberFormatType,locale:Locale? = null, configBlock:NumberFormat.()->Unit):String {
+	return getNumberFormatInstance(type,locale?:Locale.getDefault(Locale.Category.FORMAT)).apply(configBlock).format(this)
+}
+
+private fun getNumberFormatInstance(type:NumberFormatType,locale:Locale):NumberFormat {
+	return when(type){
+		NumberFormatType.Default -> NumberFormat.getInstance(locale)
+		NumberFormatType.Number -> NumberFormat.getNumberInstance(locale)
+		NumberFormatType.Integer -> NumberFormat.getIntegerInstance(locale)
+		NumberFormatType.Percent -> NumberFormat.getPercentInstance(locale)
+		NumberFormatType.Currency -> NumberFormat.getCurrencyInstance(locale)
+		else -> throw UnsupportedOperationException("Target number format type is not yet supported.")
+	}
 }